+++ CHANGELOG.md
@@ -62,3 +62,8 @@
 - add Chinese (TW) translation (@HowardZHY)
 - update config library to 1.0.13
 
+### 2.2.6
+
+- add sourcesJar to files to be uploaded to modrinth
+- add option to hide Beacon Beams
+